+++ CHANGELOG.md
@@ -7,9 +7,36 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/qutip/QuantumToolbox.jl/tree/main)
 
+## [v0.29.1]
+Release date: 2025-03-07
+
+- Minor changes for GPU matrices element type and word size handling. ([#430])
+
+## [v0.29.0]
+Release date: 2025-03-07
+
+- Add support for `OperatorKet` state input for `mesolve` and `smesolve`. ([#423])
+- Introduce `plot_fock_distribution` to plot the population of a state (ket, bra, or density matrix) in its basis (assumed to be Fock basis). ([#428])
+
+## [v0.28.0]
+Release date: 2025-02-22
+
 - Support for single `AbstractQuantumObject` in `sc_ops` for faster specific method in `ssesolve` and `smesolve`. ([#408])
 - Change save callbacks from `PresetTimeCallback` to `FunctionCallingCallback`. ([#410])
 - Align `eigenstates` and `eigenenergies` to QuTiP. ([#411])
+- Introduce `vector_to_operator` and `operator_to_vector`. ([#413])
+- Introduce some entropy related functions. ([#414], [#416])
+  - `entropy_linear`
+  - `entropy_mutual`
+  - `entropy_conditional`
+  - `entropy_relative`
+- Fix `entanglement` and introduce `concurrence`. ([#414], [#418], [#419])
+- Introduce some metric functions. ([#414], [#420])
+  - `hilbert_dist`
+  - `hellinger_dist`
+  - `bures_dist`
+  - `bures_angle`
+- Align `steadystate` ODE solver to other methods and improve GPU support. ([#421])
 
 ## [v0.27.0]
 Release date: 2025-02-14
@@ -102,6 +129,9 @@ Release date: 2024-11-13
 [v0.25.2]: https://github.com/qutip/QuantumToolbox.jl/releases/tag/v0.25.2
 [v0.26.0]: https://github.com/qutip/QuantumToolbox.jl/releases/tag/v0.26.0
 [v0.27.0]: https://github.com/qutip/QuantumToolbox.jl/releases/tag/v0.27.0
+[v0.28.0]: https://github.com/qutip/QuantumToolbox.jl/releases/tag/v0.28.0
+[v0.29.0]: https://github.com/qutip/QuantumToolbox.jl/releases/tag/v0.29.0
+[v0.29.1]: https://github.com/qutip/QuantumToolbox.jl/releases/tag/v0.29.1
 [#86]: https://github.com/qutip/QuantumToolbox.jl/issues/86
 [#139]: https://github.com/qutip/QuantumToolbox.jl/issues/139
 [#271]: https://github.com/qutip/QuantumToolbox.jl/issues/271
@@ -147,3 +177,13 @@ Release date: 2024-11-13
 [#408]: https://github.com/qutip/QuantumToolbox.jl/issues/408
 [#410]: https://github.com/qutip/QuantumToolbox.jl/issues/410
 [#411]: https://github.com/qutip/QuantumToolbox.jl/issues/411
+[#413]: https://github.com/qutip/QuantumToolbox.jl/issues/413
+[#414]: https://github.com/qutip/QuantumToolbox.jl/issues/414
+[#416]: https://github.com/qutip/QuantumToolbox.jl/issues/416
+[#418]: https://github.com/qutip/QuantumToolbox.jl/issues/418
+[#419]: https://github.com/qutip/QuantumToolbox.jl/issues/419
+[#420]: https://github.com/qutip/QuantumToolbox.jl/issues/420
+[#421]: https://github.com/qutip/QuantumToolbox.jl/issues/421
+[#423]: https://github.com/qutip/QuantumToolbox.jl/issues/423
+[#428]: https://github.com/qutip/QuantumToolbox.jl/issues/428
+[#430]: https://github.com/qutip/QuantumToolbox.jl/issues/430